requirements.txt
Package                Version
---------------------- ---------
aniso8601              9.0.1
asgiref                3.3.4
attrs                  21.2.0
bcrypt                 3.2.0
blinker                1.4
boto3                  1.17.97
botocore               1.20.97
certifi                2021.5.30
cffi                   1.14.5
chardet                4.0.0
click                  8.0.1
Django                 3.2.4
djangorestframework    3.12.4
dnspython              2.1.0
email-validator        1.1.3
Flask                  2.0.1
Flask-Bcrypt           0.7.1
Flask-Login            0.5.0
Flask-Mail             0.9.1
flask-marshmallow      0.14.0
Flask-RESTful          0.3.9
Flask-SQLAlchemy       2.5.1
Flask-WTF              0.15.1
greenlet               1.1.0
idna                   2.10
iniconfig              1.1.1
itsdangerous           2.0.1
Jinja2                 3.0.1
jmespath               0.10.0
MarkupSafe             2.0.1
marshmallow            3.12.1
marshmallow-sqlalchemy 0.26.1
packaging              20.9
pika                   1.2.0
Pillow                 8.2.0
pip                    21.1.2
pluggy                 0.13.1
py                     1.10.0
pycparser              2.20
pycryptodome           3.10.1
pyparsing              2.4.7
pytest                 6.2.4
python-dateutil        2.8.1
pytz                   2021.1
requests               2.25.1
s3transfer             0.4.2
secret                 0.8
setuptools             41.2.0
six                    1.16.0
SQLAlchemy             1.4.17
sqlparse               0.4.1
style                  1.1.0
tabulate               0.8.9
toml                   0.10.2
update                 0.0.1
urllib3                1.26.5
Werkzeug               2.0.1
WTForms                2.3.3
